Rename SkillsSearchBar to SearchBar and move into /components folder. Also, move FindSkillsModal story into /stories folder
